refactor: simplify files structure
