+++ common/internal-types/src/channels.rs
@@ -251,6 +251,24 @@ impl ChannelChange {
     }
 }
 
+/// A wrapper around [`ChannelId`] representing a Channel that is corrupted.
+#[derive(Copy, Clone, Debug, PartialEq)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+pub struct CorruptedChannelEntry(ChannelId);
+
+impl From<ChannelId> for CorruptedChannelEntry {
+    fn from(value: ChannelId) -> Self {
+        CorruptedChannelEntry(value)
+    }
+}
+
+impl CorruptedChannelEntry {
+    /// Returns the channel ID of the corrupted channel.
+    pub fn channel_id(&self) -> &ChannelId {
+        &self.0
+    }
+}
+
 /// A pair of source and destination addresses representing a channel.
 #[derive(Clone, Copy, Debug, PartialEq, Eq, Hash)]
 pub struct SrcDstPair(Address, Address);
