+++ db/entity/src/conversions/mod.rs
@@ -1,5 +1,4 @@
 pub mod account;
 pub mod channels;
-pub mod corrupted_channels;
 pub mod logs;
 pub mod tickets;
